Add ResolveRouteBinding trait
