+++ src/Illuminate/Database/Eloquent/Concerns/HasUuids.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Str;
 
 trait HasUuids
 {
+    use ResolveRouteBinding;
+
     /**
      * Initialize the trait.
      *
@@ -35,29 +37,6 @@ trait HasUuids
     public function newUniqueId()
     {
         return (string) Str::orderedUuid();
-    }
-
-    /**
-     * Retrieve the model for a bound value.
-     *
-     * @param  \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Relations\Relation  $query
-     * @param  mixed  $value
-     * @param  string|null  $field
-     * @return \Illuminate\Database\Eloquent\Relations\Relation
-     *
-     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
-     */
-    public function resolveRouteBindingQuery($query, $value, $field = null)
-    {
-        if ($field && in_array($field, $this->uniqueIds()) && ! Str::isUuid($value)) {
-            throw (new ModelNotFoundException)->setModel(get_class($this), $value);
-        }
-
-        if (! $field && in_array($this->getRouteKeyName(), $this->uniqueIds()) && ! Str::isUuid($value)) {
-            throw (new ModelNotFoundException)->setModel(get_class($this), $value);
-        }
-
-        return parent::resolveRouteBindingQuery($query, $value, $field);
     }
 
     /**
